app/requirements.txt
pwntools==4.9.0
requests==2.29.0
pymongo==4.3.3
watchdog==3.0.0
inotify==0.2.10
flask==2.3.2
flask-restful==0.3.9
flask-cors==3.0.10